split plot.hpp into separate files
